+++ CHANGELOG.md
@@ -21,6 +21,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Updated `Symfony\Yaml` to version 6.2.5
  - Updated `\ncc\Managers > PackageManager` to display more logging messages
 
+### Added
+ - Dockerfile for building a docker image of NCC
+
 ## [1.0.1] - 2023-02-07
 
 ### Added